% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{gfw_path}
\alias{gfw_path}
\title{Retrieve the GFW path, possible building a path}
\usage{
gfw_path(..., root = root_path())
}
\arguments{
\item{...}{further arguments for \code{file.path()}}

\item{root}{the root path}
}
\value{
character path description
}
\description{
Retrieve the GFW path, possible building a path
}
