% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfwdata-package.R
\docType{package}
\name{gfwdata}
\alias{gfwdata}
\alias{gfwdata-package}
\title{gfwdata package}
\description{
R language tools to manage Global Fishing Watch data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/gfwdata}
}

}
