% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{read_gfw_raster}
\alias{read_gfw_raster}
\title{Read a GFW raster file}
\usage{
read_gfw_raster(name = "distance-from-shore", path = gfw_path("raster"))
}
\arguments{
\item{name}{char, the name of the file to read}

\item{path}{char, the path to the dataset}
}
\value{
stars object
}
\description{
Read a GFW raster file
}
