% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsi.R
\name{read_mmsi_daily}
\alias{read_mmsi_daily}
\title{Read a single mmsi daily CSV file.}
\usage{
read_mmsi_daily(
  date = "2020-01-01",
  path = gfw_path("mmsi-daily-csvs-10-v2-2020"),
  form = c("raw", "sf", "stars")[1]
)
}
\arguments{
\item{date}{Date, POSIXt or (character as YYYY-mm-dd) to read}

\item{path}{char, the path to the data}

\item{form}{char, one of 'raw', 'sf' or 'stars'}
}
\value{
table (tibble), sf POINT or stars object depending upon \code{form}
}
\description{
mmsi-daily-csvs-10-v2-2020
}
