% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_root_path}
\alias{set_root_path}
\title{Set the GFW data path}
\usage{
set_root_path(
  path = "/mnt/s1/projects/ecocast/coredata/gfw",
  filename = "~/.cgfwdata"
)
}
\arguments{
\item{path}{the path that defines the location of GFW data}

\item{filename}{the name the file to store the path as a single line of text}
}
\value{
NULL invisibly
}
\description{
Set the GFW data path
}
